Move `Indirection` and `TypeTranslation` into a new module `logic`.
I plan to have more of this kind of code, and separating it from
`writer` seems wise.
